requirements.txt
pandas==2.0.3
mlflow
scikit-learn
click
pyyaml
